feat: added reference interest rate functionality ...

+++ src/main/java/pfs/lms/enquiry/referenceinterest/repository/ReferenceInterestRateRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pfs.lms.enquiry.referenceinterest.domain.ReferenceInterestRate;
 
 
-    public   interface ReferenceInterestRateRepository extends JpaRepository<ReferenceInterestRate, Integer> {
+    public   interface ReferenceInterestRateRepository extends JpaRepository<ReferenceInterestRate, Long> {
         public ReferenceInterestRate findByCode(String code);
     }
